Only try to unmount/remove mountpt if it's defined

+++ tests/sys/geom/class/uzip/conf.sh
@@ -6,8 +6,10 @@ base=`basename $0`
 
 uzip_test_cleanup()
 {
-	umount $mntpoint
-	rmdir $mntpoint
+	if [ -n "$mntpoint" ]; then
+		umount $mntpoint
+		rmdir $mntpoint
+	fi
 	geom_test_cleanup
 }
 trap uzip_test_cleanup ABRT EXIT INT TERM
